[update] airpollution viw and using Chart.js

+++ airpollution/models.py
@@ -41,7 +41,7 @@ class PollutantEntry(models.Model):
     city = models.CharField(max_length=50, default='', blank=True)
     station_code = models.CharField(max_length=30, default='', blank=True)
     station_name = models.CharField(max_length=100, default='', blank=True)
-    pollution_level = models.IntegerField()
+    pollution_level = models.FloatField()
     units = models.CharField(max_length=50, default='', blank=True)
     station_type = models.CharField(max_length=30, default='', blank=True)
     station_area = models.CharField(max_length=100, default='', blank=True)
